% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcertificate.R
\name{build_certificate}
\alias{build_certificate}
\title{Build the Certificate}
\usage{
build_certificate(names, hours, email, ass_paths1, ass_paths2, ex_dir)
}
\description{
\code{build_certificate} is the function responsible to generate the PDF.
}
