% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcertificate.R
\name{mkcertificate}
\alias{mkcertificate}
\title{Create Certificates of Attendance}
\usage{
mkcertificate(data, col_names, col_hours, col_email, col_id, ass_paths1,
  ass_paths2, ex_dir = ".")
}
\description{
\code(mkcertificate) creates certificates of attendance based on
data.
}
